PiratePICProg: Commited patches from AndThen (PIC32MX), PIC32 still needs work. Added verbose debugging print (so far buspirate and serial module). Fixed PIC24 readout support, fixed endiannes of read data (BP_PIC424Read), added correct unpacking of read data.

+++ trunk/PiratePICprog/software/pic.h
@@ -23,7 +23,7 @@ struct pic_memmap_t {
 
 struct pic_chip_t {
 	const char *name;
-	uint16_t ID;
+	uint32_t ID;
 	uint16_t family;
 
 	struct pic_memmap_t memmap[PIC_MEM_LAST];
